+++ src/rpc/mod.rs
@@ -88,6 +88,7 @@ macro_rules! for_each_rpc_method {
         $callback!($crate::rpc::chain::ForestChainExportDiff);
         $callback!($crate::rpc::chain::ForestChainExportStatus);
         $callback!($crate::rpc::chain::ForestChainExportCancel);
+        $callback!($crate::rpc::chain::ChainGetTipsetByParentState);
 
         // common vertical
         $callback!($crate::rpc::common::Session);
@@ -98,6 +99,7 @@ macro_rules! for_each_rpc_method {
         // eth vertical
         $callback!($crate::rpc::eth::EthAccounts);
         $callback!($crate::rpc::eth::EthAddressToFilecoinAddress);
+        $callback!($crate::rpc::eth::FilecoinAddressToEthAddress);
         $callback!($crate::rpc::eth::EthBlockNumber);
         $callback!($crate::rpc::eth::EthCall);
         $callback!($crate::rpc::eth::EthChainId);
@@ -585,7 +587,17 @@ where
             let filter_list = filter_list.clone();
             move |req| {
                 let is_websocket = jsonrpsee::server::ws::is_upgrade_request(&req);
-                let path = ApiPaths::from_uri(req.uri()).unwrap_or(ApiPaths::V1);
+                let path = if let Ok(p) = ApiPaths::from_uri(req.uri()) {
+                    p
+                } else {
+                    return async move {
+                        Ok(http::Response::builder()
+                            .status(http::StatusCode::NOT_FOUND)
+                            .body(Default::default())
+                            .unwrap_or_else(|_| http::Response::new(Default::default())))
+                    }
+                    .boxed();
+                };
                 let methods = methods.get(&path).cloned().unwrap_or_default();
                 let PerConnection {
                     stop_handle,