+++ src/wallet/subcommands/wallet_cmd.rs
@@ -35,7 +35,7 @@ use crate::{
 };
 use anyhow::{Context as _, bail};
 use base64::{Engine, prelude::BASE64_STANDARD};
-use clap::{Subcommand, arg};
+use clap::Subcommand;
 use dialoguer::{Password, console::Term, theme::ColorfulTheme};
 use directories::ProjectDirs;
 use num::Zero as _;
@@ -516,7 +516,8 @@ impl WalletCommands {
                         &backend.remote,
                         (message, spec, ApiTipsetKey(None)),
                     )
-                    .await?;
+                    .await?
+                    .message;
 
                     if message.gas_premium > message.gas_fee_cap {
                         anyhow::bail!("After estimation, gas premium is greater than gas fee cap")